Cleanup logs. Add DRAM validation. Move test to nightly folder
